requirements.txt
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.10.0
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyclipper==1.3.0.post6
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
